Add abstracted components
